Now downloading pgn instead of cbv format
I realised that it doesn't matter which I download as both can be opened in chessbase and copied to the twic database easily. The difference
with pgn is that I first have to convert it to cbv format in the chassbase interface. The reason for using pgn instead is because the files
can be concatinated into a single file with the cat command (Using $ cat * >twic1143-1250.pgn in terminal to be specific). It is not
possible to do this with the cbv files. This way I only have to append to the Twic database once!