Created a function which should be used for automating my code.
Now if I write a script my python program should be able to download files weekly.This function goes through a text file
"twic_number_to_be_downloaded" line by line and then calls the other function to download the file who's number is given in that line.
If this file has been downloaded successfuly we delete this number so that we don't download the same file next time. Finally, we write to the
text file the next number which should be downloaded next time.

Notes: the mode for reading and writing simultaneously in the open() function is 'r+'; the idea for deleting lines in a text file, is to go
through the lines and write each line then we do not want deleted again. We then use truncate() to delete all the old lines. Hence, only
the lines that we wrote again remain.